Move color-picker to a common directory
