+++ lib/assets/javascripts/builder/components/code-mirror/colorpicker.code-mirror.js
@@ -1,7 +1,7 @@
 var CoreView = require('backbone/core-view');
 var $ = require('jquery');
 var _ = require('underscore');
-var ColorPicker = require('builder/components/form-components/editors/fill/color-picker/color-picker.js');
+var ColorPicker = require('builder/components/color-picker/color-picker.js');
 var COLOR_KEYWORDS = require('builder/helpers/color-keywords');
 
 /**
