+++ lib/assets/javascripts/builder/components/color-picker/color-picker-template.tpl
