+++ lib/assets/javascripts/builder/components/color-picker/template.tpl
