+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/inputs/input-color-fixed.js
@@ -3,7 +3,7 @@ var $ = require('jquery');
 
 var CoreView = require('backbone/core-view');
 
-var ColorPickerView = require('builder/components/form-components/editors/fill/color-picker/color-picker');
+var ColorPickerView = require('builder/components/color-picker/color-picker');
 var TipsyTooltipView = require('builder/components/tipsy-tooltip-view');
 var InputColorTemplate = require('builder/components/form-components/editors/fill-color/inputs/input-color-fixed.tpl');
 
@@ -51,7 +51,7 @@ module.exports = CoreView.extend({
       InputColorTemplate({
         value: this._getValue(),
         opacity: this._getOpacity(),
-        help: this._help.color || ''
+        help: this._help && this._help.color || null
       })
     );
 
@@ -61,7 +61,7 @@ module.exports = CoreView.extend({
   },
 
   _initViews: function () {
-    if (this._help.color) {
+    if (this._help && this._help.color) {
       var tooltip = new TipsyTooltipView({
         el: this.$('.js-help'),
         gravity: 'w',
