+++ lib/assets/javascripts/builder/components/form-components/editors/fill/input-color/input-color-fixed-content-view.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var CoreView = require('backbone/core-view');
 var tabPaneTemplate = require('builder/components/form-components/editors/fill/fill-tab-pane.tpl');
 var createMixedLabelsTabPane = require('builder/components/tab-pane/create-mixed-labels-tab-pane');
-var ColorPicker = require('builder/components/form-components/editors/fill/color-picker/color-picker');
+var ColorPicker = require('builder/components/color-picker/color-picker');
 var InputColorFileView = require('./input-color-file-view');
 var FillConstants = require('builder/components/form-components/_constants/_fill');
