+++ lib/assets/javascripts/builder/components/form-components/editors/fill/input-color/input-color-picker/input-color-picker-view.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
-var ColorPicker = require('builder/components/form-components/editors/fill/color-picker/color-picker');
+var ColorPicker = require('builder/components/color-picker/color-picker');
 var template = require('./input-color-picker-view.tpl');
 var InputColorPickerHeader = require('./input-color-picker-header');
 var checkAndBuildOpts = require('builder/helpers/required-opts');
