+++ lib/assets/javascripts/builder/components/form-components/editors/stroke-color/inputs/input-stroke-color.js
@@ -3,7 +3,7 @@ var $ = require('jquery');
 
 var CoreView = require('backbone/core-view');
 
-var ColorPickerView = require('builder/components/form-components/editors/fill/color-picker/color-picker');
+var ColorPickerView = require('builder/components/color-picker/color-picker');
 var TipsyTooltipView = require('builder/components/tipsy-tooltip-view');
 var InputStrokeColorTemplate = require('builder/components/form-components/editors/stroke-color/inputs/input-stroke-color.tpl');
